% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_npp.R
\name{glm_npp}
\alias{glm_npp}
\title{Normalized power prior for generalized linear models using Stan}
\usage{
glm_npp(
  formula,
  family,
  data,
  histdata,
  a0.lognc,
  beta0 = NULL,
  Sigma0 = NULL,
  offset = NULL,
  disp.shape = 1.5,
  disp.scale = 0.25,
  a0.shape1 = 1,
  a0.shape2 = 1,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}.}

\item{family}{an object of type \code{\link[stats]{family}} giving distribution and link function}

\item{data}{a \code{\link[base]{data.frame}} of current data giving all variables in \code{formula}}

\item{histdata}{a \code{\link[base]{data.frame}} of historical data giving all variables in \code{formula}}

\item{a0.lognc}{a \code{matrix} whose first column is a value for the power prior and second column gives the estimated log normalizing constant}

\item{beta0}{mean for initial prior on regression coefficients. Defaults to vector of 0s}

\item{Sigma0}{covariance matrix for initial prior on regression coefficients. Defaults to \code{diag(100, ncol(X))}}

\item{offset}{offset in GLM. If \code{NULL}, no offset is utilized}

\item{disp.shape}{shape parameter for inverse-gamma prior on dispersion parameter (for Gaussian and gamma models). Ignored for binomial and Poisson models}

\item{disp.scale}{rate parameter for inverse-gamma prior on dispersion parameter (for Gaussian and gamma models). Ignored for binomial and Poisson models}

\item{a0.shape1}{shape1 parameter for beta prior on the power prior parameter. When \code{a0.shape1==1} and \code{a0.shape2==1}, a uniform prior is utilized}

\item{a0.shape2}{shape2 parameter for beta prior on the power prior parameter. When \code{a0.shape1==1} and \code{a0.shape2==1}, a uniform prior is utilized}

\item{...}{optional parameters to pass onto \code{rstan::sampling}}
}
\value{
an object of class \link[rstan:stanfit-class]{rstan::stanfit} returned by \code{rstan::sampling}
}
\description{
Sample from the posterior of a normalized power prior (NPP) generalized linear model (GLM).
Requires that the log normalizing constant of the prior was estimated in a grid of values between 0 and 1.
}
\examples{
N = 50
}
