% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_npp_prior.R
\name{glm_npp_prior}
\alias{glm_npp_prior}
\title{Sample from prior of power prior for a GLM}
\usage{
glm_npp_prior(
  formula,
  family,
  histdata,
  a0,
  beta0 = NULL,
  Sigma0 = NULL,
  offset = NULL,
  disp.shape = 1e-04,
  disp.scale = 1e-04,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}.}

\item{family}{an object of type \code{\link[stats]{family}} giving distribution and link function}

\item{histdata}{a \code{\link[base]{data.frame}} of historical data giving all variables in \code{formula}}

\item{a0}{a positive scalar no larger than 1 giving the power prior parameter}

\item{beta0}{mean for initial prior on regression coefficients. Defaults to vector of 0s}

\item{Sigma0}{covariance matrix for initial prior on regression coefficients. Defaults to \code{diag(100, ncol(X))}}

\item{offset}{offset in GLM. If \code{NULL}, no offset is utilized}

\item{disp.shape}{shape parameter for inverse-gamma prior on dispersion (for Gaussian and gamma models). Ignored for binomial and Poisson models}

\item{disp.scale}{scale parameter for inverse-gamma prior on dispersion (for Gaussian and gamma models). Ignored for binomial and Poisson models}

\item{...}{optional parameters to pass onto \code{rstan::sampling}}
}
\value{
an object of class \link[rstan:stanfit-class]{rstan::stanfit} returned by \code{rstan::sampling}
}
\description{
Uses \code{rstan} to sample from the prior of a power prior for a generalized linear model
}
\examples{
N = 50
}
