% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanglm.R
\name{rstanglm}
\alias{rstanglm}
\title{Bayesian linear regression with conjugate prior}
\usage{
rstanglm(
  formula,
  family,
  data,
  prior.precision = 1,
  prior.y0,
  dispersion,
  dispersion.shape = NULL,
  dispersion.rate = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}.
Two-sided formulas (e.g., \code{y ~ x}) samples from the posterior density while
one-sided formulas (e.g., \code{~ x}) samples from the prior density.}

\item{family}{an object of type \code{\link[stats]{family}} giving distribution and link function}

\item{data}{a \code{\link[base]{data.frame}} giving all variables in \code{formula}}

\item{prior.precision}{a scalar giving the prior precision parameter. This is denoted by \eqn{a_0} in Chen and Ibrahim (2003)}

\item{prior.y0}{a vector of length \code{nrow(data)} giving the pseudo-data.
The range of each element of \code{prior.y0} should be same as the mean (e.g., for Bernoulli models, each element should
be between 0 and 1). This is denoted by \eqn{y_0} in Chen and Ibrahim (2003)}

\item{dispersion}{a scalar or string. If a scalar, assumes fixed dispersion parameter equal to a scalar. If a string, must be one of 'mle' or
random. If 'random' is specified or formula, the user must supply positive
scalar parameter values to \code{dispersion.shape} and \code{dispersion.rate}. Must be a positive scalar if
\code{formula} is one-sided and family is not \code{binomial} or \code{poisson}.}

\item{dispersion.shape}{positive scalar giving prior shape hyperparameter for dispersion parameter (if \code{dispersion=="random"}). Ignored otherwise.}

\item{dispersion.rate}{positive scalar giving prior rate hyperparameter for dispersion parameter (if \code{dispersion=="random"}). Ignored otherwise.}

\item{...}{arguments passed to \code{\link[rstan]{sampling}} (e.g. iter, chains).}
}
\value{
an object of class \link[rstan:stanfit-class]{rstan::stanfit} returned by \code{rstan::sampling}
}
\description{
Fits a Bayesian generalized linear model using the conjugate prior of Chen and Ibrahim (2003).
}
\examples{
N = 50
x = rnorm(N, sd = 5)
y = rnorm(0.5 + 0.25 * x, sd = 1)
data = data.frame(y = y, x = x)
## NOT RUN--SAMPLE FROM PRIOR
# prior_sample = rstanglm_fixedDispersion(
#   ~ x, gaussian(), data = data,
#   prior.y0 = rep(0, N), dispersion = 1)

## SAMPLE FROM POSTERIOR
# options(mc.cores = parallel::detectCores())
rstan::rstan_options(auto_write = TRUE)
post_sample = rstanglm(
y ~ x, gaussian(), data = data, prior.y0 = rep(0, N),
dispersion = 'random', dispersion.shape = 1, dispersion.rate = .1)  ## sample from posterior with random dispersion parameter

}
