% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanpp-package.R
\docType{package}
\name{rstanpp-package}
\alias{rstanpp-package}
\alias{rstanpp-pkg}
\title{The 'rstanpp' package.}
\description{
Normalized power prior analysis using Stan
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
